+++ src/gallery-itsawidgetrenderpromise/HISTORY.md
@@ -1,9 +1,16 @@
 gallery-itsawidgetrenderpromise
 ========
 
+0.3
+-----
+
+* Added Y.Widget.renderWhenAvailable()
+* Added Y.Widget.renderWhenAvailablePromise()
+
 0.2
 -----
 
+* Added Y.Widget.renderOnAvailable()
 * Added Y.Widget.renderOnAvailablePromise()
 
 0.1